[ADD][FLASK] add editpage with password
