+++ FLASK/hstack/hstack/static/css/videoDetail.css
@@ -629,3 +629,11 @@ div.right {
     border-radius: 10px;
     overflow: hidden;
 }
+
+.right {
+    height: 100%;
+    display: flex;
+    align-items: flex-end;
+    justify-content: flex-end;
+    padding: 20px 20px;
+}
